+++ lazy.lua
@@ -1,0 +1,5 @@
+return {
+    "catboylei/bfDisplay-rs",
+    build = "bash lazy-install.sh",
+    opts = {}
+}