+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Fixed
 - Pointer off-by-one error in cell display
+- Mousescroll not being properly set on cell display
 
 ## [1.0.0] - 2026-03-10
 - Initial release